Update rollJam.ino
I just discovered samy's rolljam idea and I found your code, I got interested and bought a teensy 3.2 and two cc1101 to try to use your code and understand how it works, while reading it I found out that the way the button operates could be changed to avoid buying a resister, that's a whole 10 cents cheaper! 😛 
Other than that can't figure out why it's picking up random packets, I just power it on and in less than a minutes it picks up at least a packet but i'm sure it's not in the 433mhz range...
Could it be a problem with my cc1101 or I'm just dumb?